+++ CHANGELOG.md
@@ -2458,7 +2458,7 @@
 
 ### Features
 
-* **language:** one shift(over=, offset=, edge=), replacing roll and shift ([#359](https://github.com/fluxopt/lpspec/issues/359)) ([8473a24](https://github.com/fluxopt/lpspec/commit/8473a24621326eb39151fd50337f1c6decb7a51d))
+* **language:** one shift(along=, offset=, edge=), replacing roll and shift ([#359](https://github.com/fluxopt/lpspec/issues/359)) ([8473a24](https://github.com/fluxopt/lpspec/commit/8473a24621326eb39151fd50337f1c6decb7a51d))
 
 ## [0.0.1-alpha.34](https://github.com/fluxopt/lpspec/compare/v0.0.1-alpha.33...v0.0.1-alpha.34) (2026-07-30)